requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-docx==1.1.2
PyPDF2==3.0.1
pytesseract==0.3.13
Pillow==10.4.0
groq
celery[redis]==5.4.0
redis==5.0.8
python-dotenv==1.0.1
pydantic==2.9.2
